Fix telemetry use by the apps-web repository
Noticed a discrepency where the latest telemetry wasn't being applied to the app cards in the home, catalog, or individual app pages.

Investigated and found this is caused by a conflicting cache problem with the telemetry pull and usage in the various shortcodes.

Cleaned up these shortcodes to use a unified approach to telemetry and avoid cache issues.

Local build testing has been good.